+++ src/main/javaLib/novah/range/IntRange.java
@@ -58,7 +58,7 @@ public class IntRange implements Range<Integer> {
     }
 
     /**
-     * Creates a iterator that will go through all numbers in this range
+     * Creates an iterator that will go through all numbers in this range
      */
     @Override
     public Iterator<Integer> iterator() {